move data test in full-sync suite
